Move content of db to ext
